[JIT] Add JIT support for torch.no_grad
**Summary**
This commit enables the use of `torch.no_grad()` in a with item of a
with statement within JIT. Note that the use of this context manager as
a decorator is not supported.

**Test Plan**
This commit adds a test case to the existing with statements tests for
`torch.no_grad()`.

**Fixes**
This commit fixes #40259.

ghstack-source-id: faada1e83e9942103be37af6affe8016638d0015
Pull Request resolved: https://github.com/pytorch/pytorch/pull/41371